pyproject.toml
[project]
name = "google-adk-commentator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.6.15",
    "google-adk>=1.3.0",
    "litellm>=1.72.6",
    "loguru>=0.7.3",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pygame>=2.6.1",
]
